Bogos

+++ Assets/HealthBarScript.cs
@@ -5,7 +5,7 @@ using UnityEngine.UI;
 
 public class HealthBarScript : MonoBehaviour
 {
-    public void changeHealth(int health)
+    public void changeHealth(float health)
     {
         Slider slider = GetComponent<Slider>();
         slider.value = health;
